requirements.txt
streamlit>=1.28.0
pandas>=1.3.0
numpy>=1.20.0
plotly>=5.13.0
pydantic>=1.10.0
pydantic-settings>=2.0.0
google-cloud-bigquery>=3.9.0
google-api-python-client>=2.0.0
supabase>=2.0.0
pandas-gbq>=0.19.0
google-auth>=2.0.0
requests>=2.25.0
pyarrow>=12.0.0
db-dtypes>=1.1.1
altair>=5.0.0
matplotlib>=3.5.0